refactor server
